+++ instrumentation/opentelemetry-instrumentation-falcon/tests/test_falcon.py
@@ -120,6 +120,9 @@ _recommended_metrics_attrs_both = {
 _parsed_falcon_version = package_version.parse(_falcon_version)
 
 
+SCOPE = "opentelemetry.instrumentation.falcon"
+
+
 class TestFalconBase(TestBase):
     def setUp(self):
         super().setUp()
@@ -487,7 +490,7 @@ class TestFalconInstrumentation(TestFalconBase, WsgiTestBase):
             self.assertFalse(mock_span.set_attribute.called)
             self.assertFalse(mock_span.set_status.called)
 
-            metrics = self.get_sorted_metrics()
+            metrics = self.get_sorted_metrics(SCOPE)
             self.assertTrue(len(metrics) != 0)
             for metric in metrics:
                 data_points = list(metric.data.data_points)
@@ -519,7 +522,7 @@ class TestFalconInstrumentation(TestFalconBase, WsgiTestBase):
         self.client().simulate_get("/hello/756")
         self.client().simulate_get("/hello/756")
         self.client().simulate_get("/hello/756")
-        metrics = self.get_sorted_metrics()
+        metrics = self.get_sorted_metrics(SCOPE)
         number_data_point_seen = False
         histogram_data_point_seen = False
         self.assertTrue(len(metrics) != 0)
@@ -545,7 +548,7 @@ class TestFalconInstrumentation(TestFalconBase, WsgiTestBase):
         self.client().simulate_get("/hello/756")
         duration = max(default_timer() - start, 0)
 
-        metrics = self.get_sorted_metrics()
+        metrics = self.get_sorted_metrics(SCOPE)
         for metric in metrics:
             data_points = list(metric.data.data_points)
             self.assertEqual(len(data_points), 1)
@@ -577,7 +580,7 @@ class TestFalconInstrumentation(TestFalconBase, WsgiTestBase):
         self.client().simulate_get("/hello/756")
         duration_s = default_timer() - start
 
-        metrics = self.get_sorted_metrics()
+        metrics = self.get_sorted_metrics(SCOPE)
 
         # pylint: disable=too-many-nested-blocks
         for metric in metrics:
@@ -626,7 +629,7 @@ class TestFalconInstrumentation(TestFalconBase, WsgiTestBase):
         self.client().simulate_get("/hello/756")
         duration = max(round((default_timer() - start) * 1000), 0)
 
-        metrics = self.get_sorted_metrics()
+        metrics = self.get_sorted_metrics(SCOPE)
         for metric in metrics:
             data_points = list(metric.data.data_points)
             self.assertEqual(len(data_points), 1)
@@ -650,7 +653,7 @@ class TestFalconInstrumentation(TestFalconBase, WsgiTestBase):
         self.client().simulate_request(method="POST", path="/hello/756")
         FalconInstrumentor().uninstrument()
         self.client().simulate_request(method="POST", path="/hello/756")
-        metrics = self.get_sorted_metrics()
+        metrics = self.get_sorted_metrics(SCOPE)
         for metric in metrics:
             for point in list(metric.data.data_points):
                 if isinstance(point, HistogramDataPoint):